// src/types/types.tsx
export interface CardFields {
    picture: string
    minerId: number
    claimPassId: number
}

export interface TraitsFields {
    Background: string[]
    Backpack: string[]
    Body: string[]
    Clothing: string[]
    Eyes: string[]
    Face: string[]
    Hair: string[]
    "Facial Hair": string[]
    Head: string[]
    Headlamp: string[]

}